Fold JupiterTestSuite into JupiterTestExtension
Reduces the JUnit5-migration PR's new-file count from 4 to 3. JupiterTestSuite
has no cross-package callers (only ModelTestSuite, same package), so it
becomes a package-private static nested class inside JupiterTestExtension,
the type that does have cross-package callers (@ExtendWith(JupiterTestExtension)
in plugins/example and any future plugin). No public API or behavior change.
